requirements.txt
discord.py==1.7.3
spotipy==2.19.0
python-dotenv==0.19.2
aiohttp==3.8.1
